Revert move struct defs - use forward decl instead
